feat(sbb-paginator): move to new folder and create common files
