+++ src/elements/paginator.ts
@@ -1,1 +1,2 @@
+export * from './paginator/common.js';
 export * from './paginator/paginator.js';
