+++ src/elements/paginator/common.ts
@@ -1,0 +1,1 @@
+export * from './common/paginator-common.js';
