--- src/elements/paginator/index.ts
@@ -1,1 +1,0 @@
-export * from './paginator.js';
