+++ src/elements/paginator/paginator/paginator.ssr.spec.ts
@@ -1,7 +1,7 @@
 import { assert } from '@open-wc/testing';
 import { html } from 'lit/static-html.js';
 
-import { ssrHydratedFixture } from '../core/testing/private.js';
+import { ssrHydratedFixture } from '../../core/testing/private.js';
 
 import { SbbPaginatorElement } from './paginator.js';
