+++ src/elements/paginator/paginator/paginator.stories.ts
@@ -11,7 +11,7 @@ import type {
 import type { TemplateResult } from 'lit';
 import { html } from 'lit';
 
-import { sbbSpread } from '../../storybook/helpers/spread.js';
+import { sbbSpread } from '../../../storybook/helpers/spread.js';
 
 import { SbbPaginatorElement } from './paginator.js';
 import readme from './readme.md?raw';
