+++ src/elements/paginator/paginator/readme.md
@@ -57,15 +57,15 @@ that describes the content controlled by the paginator.
 
 ## Properties
 
-| Name              | Attribute           | Privacy | Type                    | Default   | Description                                                                                                                                                                      |
-| ----------------- | ------------------- | ------- | ----------------------- | --------- | -------------------------------------------------------------------------------------------------------------------------------------------------------------------------------- |
-| `length`          | `length`            | public  | `number`                | `0`       | Total number of items.                                                                                                                                                           |
-| `negative`        | `negative`          | public  | `boolean`               | `false`   | Negative coloring variant flag.                                                                                                                                                  |
-| `pageIndex`       | `page-index`        | public  | `number`                | `0`       | Current page index.                                                                                                                                                              |
-| `pagerPosition`   | `pager-position`    | public  | `'start' \| 'end'`      | `'start'` | Position of the prev/next buttons: if `pageSizeOptions` is set, the sbb-select for the pageSize change will be positioned oppositely with the page numbers always in the center. |
-| `pageSize`        | `page-size`         | public  | `number`                | `10`      | Number of items per page.                                                                                                                                                        |
-| `pageSizeOptions` | `page-size-options` | public  | `number[] \| undefined` |           | The available `pageSize` choices.                                                                                                                                                |
-| `size`            | `size`              | public  | `'m' \| 's'`            | `'m'`     | Size variant, either m or s.                                                                                                                                                     |
+| Name              | Attribute           | Privacy | Type                        | Default   | Description                                                                                                                                                                      |
+| ----------------- | ------------------- | ------- | --------------------------- | --------- | -------------------------------------------------------------------------------------------------------------------------------------------------------------------------------- |
+| `length`          | `length`            | public  | `number`                    | `0`       | Total number of items.                                                                                                                                                           |
+| `negative`        | `negative`          | public  | `boolean`                   | `false`   | Negative coloring variant flag.                                                                                                                                                  |
+| `pageIndex`       | `page-index`        | public  | `number`                    | `0`       | Current page index.                                                                                                                                                              |
+| `pagerPosition`   | `pager-position`    | public  | `\| 'start'       \| 'end'` | `'start'` | Position of the prev/next buttons: if `pageSizeOptions` is set, the sbb-select for the pageSize change will be positioned oppositely with the page numbers always in the center. |
+| `pageSize`        | `page-size`         | public  | `number`                    | `10`      | Number of items per page.                                                                                                                                                        |
+| `pageSizeOptions` | `page-size-options` | public  | `number[] \| undefined`     |           | The available `pageSize` choices.                                                                                                                                                |
+| `size`            | `size`              | public  | `'m' \| 's'`                | `'m'`     | Size variant, either m or s.                                                                                                                                                     |
 
 ## Events
